requirements.txt
typer==0.12.3
rich==13.7.1
pyyaml==6.0.2
fastapi==0.111.1
uvicorn==0.30.6
requests==2.32.3
grpcio==1.65.5
grpcio-tools==1.65.5
pytest==8.3.2
hypothesis==6.112.1
jsonschema==4.23.0
openapi-schema-validator
